build: Add meson build
phosphor-ipmi-flash: Fix all warning for Meson build

TODO: Need to fix the issue with `alignof` on non-type in memcpyAligned.

Changes to note,
- `with_systemdsystemunitdir` and `with_tmpfilesdir` is removed since it
  is not being documented and used in OpenBMC.
- Removed Code coverage with `-DDHAVE_GCOV` does not use it and meson
  covers it.
- Removed `--enable-oe-sdk` for using the OpenBMC SDK. It should with
  directly with no change required.

+++ bmc/firmware-handler/test/firmware_multiplebundle_unittest.cpp
@@ -109,8 +109,8 @@ class IpmiOnlyTwoFirmwaresTest : public ::testing::Test
     TriggerMock *biosPrepareMockPtr, *biosVerifyMockPtr, *biosUpdateMockPtr;
 
     std::uint16_t session = 1;
-    std::uint16_t flags =
-        blobs::OpenFlags::write | FirmwareFlags::UpdateFlags::ipmi;
+    std::uint16_t flags = static_cast<std::uint16_t>(blobs::OpenFlags::write) |
+                          FirmwareFlags::UpdateFlags::ipmi;
 };
 
 TEST_F(IpmiOnlyTwoFirmwaresTest, OpeningBiosAfterBlobFails)
